Feature: scihub-1209 google slide modal (#87)
* feature/SCIHUB-1209: fixed typo in hook name

* feature/SCIHUB-1209 add component prop

* feature/SCIHUB-1209: fixed naming

* feature/SCIHUB-1209: increment version

* feature/SCIHUB-1209: whole modal as prop

---------
